+++ Runtime/Modules/Resource/Resource/ResourceService.Records.cs
@@ -31,13 +31,13 @@ namespace AlicizaX.Resource.Runtime
         private int _lastIdleProcessTick;
         private readonly ResourceUlongIntMap _assetRecordsByKey = new ResourceUlongIntMap();
         private readonly ResourceIndexMap<int, int> _assetRecordByLoadKeyId = new ResourceIndexMap<int, int>();
-        private readonly ResourceIndexMap<int, int> _assetRecordHeadByUnityObjectId = new ResourceIndexMap<int, int>();
+        private readonly ResourceIndexMap<ulong, int> _assetRecordHeadByUnityObjectId = new ResourceIndexMap<ulong, int>();
 
         private struct AssetSlot
         {
             public ulong Key;
             public int LoadKeyId;
-            public int AssetInstanceId;
+            public ulong AssetInstanceId;
             public Object Asset;
             public AssetHandle AssetHandle;
             public SubAssetsHandle SubAssetsHandle;
@@ -990,7 +990,7 @@ namespace AlicizaX.Resource.Runtime
                 return false;
             }
 
-            int instanceId = UnityObjectId.Get(asset);
+            ulong instanceId = UnityObjectId.Get(asset);
             if (!_assetRecordHeadByUnityObjectId.TryGetValue(instanceId, out int current))
             {
                 return false;
@@ -1035,7 +1035,7 @@ namespace AlicizaX.Resource.Runtime
                 return false;
             }
 
-            int instanceId = UnityObjectId.Get(unityObject);
+            ulong instanceId = UnityObjectId.Get(unityObject);
             if (!_assetRecordHeadByUnityObjectId.TryGetValue(instanceId, out int current))
             {
                 return false;
@@ -1471,7 +1471,7 @@ namespace AlicizaX.Resource.Runtime
 
         private void UnlinkAssetByUnityObject(int assetId, ref AssetSlot slot)
         {
-            int instanceId = slot.AssetInstanceId;
+            ulong instanceId = slot.AssetInstanceId;
             if (instanceId == 0 || !_assetRecordHeadByUnityObjectId.TryGetValue(instanceId, out int current))
             {
                 return;
